adapter: introduce TimestampOracle trait, move durable oracle to separate mod
