+++ src/adapter/src/coord/message_handler.rs
@@ -29,6 +29,7 @@ use tracing::{event, warn, Instrument, Level};
 use crate::client::ConnectionId;
 use crate::command::{Command, ExecuteResponse};
 use crate::coord::appends::Deferred;
+use crate::coord::timestamp_oracle::TimestampOracle;
 use crate::coord::{
     Coordinator, CreateConnectionValidationReady, Message, PeekStage, PeekStageFinish,
     PendingReadTxn, PlanValidity, PurifiedStatementReady, RealTimeRecencyContext,
